Changed the order in which we search the patients when posting a journal entry

+++ src/main/java/com/uade/psyline/application/usecase/PatientUsecase.java
@@ -16,6 +16,7 @@ import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -69,33 +70,36 @@ public class PatientUsecase implements PatientService {
         JournalEntryDAO newJournalEntryDAO = mapper.map(newJournalEntryDTO, JournalEntryDAO.class);
         newJournalEntryDAO.setPatient(patientFoundDAO);
         journalEntryRepository.save(newJournalEntryDAO);
-        PatientDTO patientDTO = mapper.map(patientFoundDAO, PatientDTO.class);
-
-        return patientDTO;
+        return  mapper.map(patientFoundDAO, PatientDTO.class);
     }
 
     @Override
     @Transactional
     public PatientDTO updateJournalEntry(String userName, JournalEntryDTO updatedJournalEntryDTO) {
         JournalEntryDAO existingJournalEntryDAO = this.findJournalEntryBy(updatedJournalEntryDTO.getId());
+        PatientDAO patientFoundDAO = this.findPatientByUserName(userName);
+        if(existingJournalEntryDAO.getPatient() != patientFoundDAO){
+            throw new JournalEntryNotFoundException();
+        }
         JournalEntryDAO updatedJournalEntryDAo = mapper.map(updatedJournalEntryDTO, JournalEntryDAO.class);
         mapper.getConfiguration().setMatchingStrategy(MatchingStrategies.STRICT);
         mapper.getConfiguration().setPropertyCondition(Conditions.isNotNull());
         mapper.map(updatedJournalEntryDAo, existingJournalEntryDAO);
         journalEntryRepository.save(existingJournalEntryDAO);
 
-        PatientDAO patientFoundDAO = this.findPatientByUserName(userName);
+
         return mapper.map(patientFoundDAO, PatientDTO.class);
     }
 
     @Override
-    @Transactional
     public PatientDTO deleteJournalEntry(String userName, Integer journalEntryId) {
         JournalEntryDAO existingJournalEntryDAO = this.findJournalEntryBy(journalEntryId);
-        journalEntryRepository.delete(existingJournalEntryDAO);
-
         PatientDAO patientFoundDAO = this.findPatientByUserName(userName);
-        return mapper.map(patientFoundDAO, PatientDTO.class);
+        if (!Objects.equals(existingJournalEntryDAO.getPatient().getUserName(), patientFoundDAO.getUserName())) {
+            throw new JournalEntryNotFoundException();
+        }
+        journalEntryRepository.delete(existingJournalEntryDAO);
+        return mapper.map(this.findPatientByUserName(userName), PatientDTO.class);
     }
 
     private PatientDAO findPatientByUserName(String userName) {
